+++ R/ggpredict.R
@@ -80,15 +80,14 @@ ggpredict <- function(model,
   ## TODO: remove deprecated later
   if (!missing(ppd) && isTRUE(ppd)) {
     insight::format_warning("Argument `ppd` is deprecated and will be removed in the future. Please use `interval` instead.") # nolint
+    interval <- "prediction"
   }
 
   # check arguments
-  type_and_ppd <- .validate_type_argument(model, type, ppd)
-  type <- type_and_ppd$type
-  ppd <- type_and_ppd$ppd
+  type <- .validate_type_argument(model, type)
 
   if (missing(interval)) {
-    if (type %in% c("re", "re.zi")) {
+    if (type %in% c("random", "zero_inflated_random")) {
       interval <- "prediction"
     } else {
       interval <- "confidence"
@@ -99,13 +98,6 @@ ggpredict <- function(model,
 
   # make sure we have valid values
   interval <- match.arg(interval, c("confidence", "prediction"))
-
-  ## TODO: remove when deprecated
-
-  # update interval - if we have ppd = TRUE, we have prediction intervals
-  if (isTRUE(ppd)) {
-    interval <- "prediction"
-  }
 
   ## TODO: remove deprecated later
 
@@ -234,7 +226,7 @@ ggpredict_helper <- function(model,
   model_info <- .get_model_info(model)
 
   # survival models are binomial
-  if (model_class == "coxph" && type == "surv") {
+  if (model_class == "coxph" && type == "survival") {
     model_info$is_binomial <- TRUE
   }
 
@@ -242,7 +234,7 @@ ggpredict_helper <- function(model,
   # done for random effects only (i.e. all focal terms are specified as random
   # effects in the model). If so, we need to tell the user that they should
   # better to `margin = "empirical"`
-  if (!type %in% c("re", "re.zi")) {
+  if (!type %in% c("random", "zero_inflated_random")) {
     .check_focal_for_random(model, terms, verbose)
   }
 
@@ -291,7 +283,7 @@ ggpredict_helper <- function(model,
 
   # for survival probabilities or cumulative hazards, we need
   # the "time" variable
-  if (model_class == "coxph" && type %in% c("surv", "cumhaz")) {
+  if (model_class == "coxph" && type %in% c("survival", "cumulative_hazard")) {
     terms <- c("time", terms)
     cleaned_terms <- c("time", cleaned_terms)
   }
@@ -326,7 +318,7 @@ ggpredict_helper <- function(model,
   )
 
   # no adjustment for type = "simulate"
-  if (type == "sim") {
+  if (type == "simulate") {
     attributes(data_grid)$constant.values <- NULL
   }